+++ CHANGELOG.md
@@ -134,7 +134,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
     - `plate_width: Optional[float] = None`
     - `use_unsafe_hotel: bool = False`
     - `iswap_collision_control_level: int = 0`
-    - `iswap_fold_up_sequence_at_the_end_of_process: bool = True`
+    - `iswap_fold_up_sequence_at_the_end_of_process: bool = False`
   - STAR.drop_resource
     - `minimum_traverse_height_at_beginning_of_a_command: Optional[float] = None`
     - `z_position_at_the_command_end: Optional[float] = None`